added password policy to the backend
